requirements_local.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
gast==0.2.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.9.0
idna==2.9
importlib-metadata==1.6.1
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
nltk==3.5
numpy==1.19.0
Keras==2.2.5
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
sklearn==0.0
tensorboard==2.0.2
tensorflow-gpu==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.9
Werkzeug==1.0.1
keras_bert==0.81.0
keras-transformer==0.31.0
wrapt==1.12.1
zipp==3.1.0
keras-multi-head==0.27.0
gensim==3.8.3
matplotlib==3.2.2
tensorflow-hub==0.7.0
